pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sentinel"
version = "0.1.0"
description = "An LLM-powered conversational agent with persistent, timestamped memory powered by a knowledge graph."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Paul" }]
dependencies = [
  "camel-ai",
  "unstructured",
  "neo4j",
  "python-dotenv",
]

[project.scripts]
sentinel = "sentinel.main:main"

[tool.setuptools]
packages = ["sentinel"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]